backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
faiss-cpu
textstat
jinja2
python-dotenv
openai
anthropic
requests
